MB-20850. Add TOKENS() function to enumerate all tokens in a value.
Reviewed-on: http://review.couchbase.org/67662
Tested-by: Gerald Sangudi <gerald@couchbase.com>

+++ expression/func_obj.go
@@ -384,7 +384,7 @@ func (this *ObjectLength) Apply(context Context, arg value.Value) (value.Value, 
 	}
 
 	oa := arg.Actual().(map[string]interface{})
-	return value.NewValue(float64(len(oa))), nil
+	return value.NewValue(len(oa)), nil
 }
 
 /*
